moving kml stuff to kml
